+++ packages/react-components/react-button/src/components/ToggleButton/useToggleButtonStyles.styles.ts
@@ -62,7 +62,7 @@ const useRootCheckedStyles = makeStyles({
       },
 
       ':focus': {
-        ...shorthands.border('1px', 'solid', 'HighlightText'),
+        border: '1px solid HighlightText',
         outlineColor: 'Highlight',
       },
     },