+++ packages/react-components/react-motions-preview/src/factories/createPresenceComponent.ts
@@ -1,17 +1,13 @@
-import { useEventCallback, useIsomorphicLayoutEffect, useMergedRefs } from '@fluentui/react-utilities';
-import type { EventData, EventHandler } from '@fluentui/react-utilities';
+import { useEventCallback, useFirstMount, useIsomorphicLayoutEffect, useMergedRefs } from '@fluentui/react-utilities';
 import * as React from 'react';
 
 import { PresenceGroupChildContext } from '../contexts/PresenceGroupChildContext';
 import { useIsReducedMotion } from '../hooks/useIsReducedMotion';
 import { useMotionImperativeRef } from '../hooks/useMotionImperativeRef';
 import { useMountedState } from '../hooks/useMountedState';
+import { animate } from '../utils/animate';
 import { getChildElement } from '../utils/getChildElement';
 import type { PresenceMotion, MotionImperativeRef, PresenceMotionFn } from '../types';
-
-type PresenceMotionEventData = EventData<'animation', AnimationPlaybackEvent> & {
-  direction: 'enter' | 'exit';
-};
 
 export type PresenceComponentProps = {
   /**
@@ -26,7 +22,14 @@ export type PresenceComponentProps = {
   /** Provides imperative controls for the animation. */
   imperativeRef?: React.Ref<MotionImperativeRef | undefined>;
 
-  onMotionFinish?: EventHandler<PresenceMotionEventData>;
+  /**
+   * Callback that is called when the whole motion finishes.
+   *
+   * A motion definition can contain multiple animations and therefore multiple "finish" events. The callback is
+   * triggered once all animations have finished with "null" instead of an event object to avoid ambiguity.
+   */
+  // eslint-disable-next-line @nx/workspace-consistent-callback-type -- EventHandler<T> does not support "null"
+  onMotionFinish?: (ev: null, data: { direction: 'enter' | 'exit' }) => void;
 
   /** Defines whether a component is visible; triggers the "enter" or "exit" motions. */
   visible?: boolean;
@@ -55,14 +58,14 @@ export function createPresenceComponent(motion: PresenceMotion | PresenceMotionF
     const ref = useMergedRefs(elementRef, child.ref);
     const optionsRef = React.useRef<{ appear?: boolean }>({});
 
-    const isFirstMount = React.useRef<boolean>(true);
+    const isFirstMount = useFirstMount();
     const isReducedMotion = useIsReducedMotion();
 
-    const onEnterFinish = useEventCallback((event: AnimationPlaybackEvent) => {
-      onMotionFinish?.(event, { event, type: 'animation', direction: 'enter' });
+    const onEnterFinish = useEventCallback(() => {
+      onMotionFinish?.(null, { direction: 'enter' });
     });
-    const onExitFinish = useEventCallback((event: AnimationPlaybackEvent) => {
-      onMotionFinish?.(event, { event, type: 'animation', direction: 'exit' });
+    const onExitFinish = useEventCallback(() => {
+      onMotionFinish?.(null, { direction: 'exit' });
 
       if (unmountOnExit) {
         setMounted(false);
@@ -74,39 +77,44 @@ export function createPresenceComponent(motion: PresenceMotion | PresenceMotionF
       optionsRef.current = { appear };
     });
 
-    useIsomorphicLayoutEffect(() => {
-      if (!elementRef.current || shouldSkipAnimation(optionsRef.current.appear, isFirstMount.current, visible)) {
-        return;
-      }
+    useIsomorphicLayoutEffect(
+      () => {
+        if (!elementRef.current || shouldSkipAnimation(optionsRef.current.appear, isFirstMount, visible)) {
+          return;
+        }
 
-      const presenceDefinition = typeof motion === 'function' ? motion(elementRef.current) : motion;
-      const { keyframes, ...options } = visible ? presenceDefinition.enter : presenceDefinition.exit;
+        const presenceDefinition = typeof motion === 'function' ? motion(elementRef.current) : motion;
+        const { keyframes, ...options } = visible ? presenceDefinition.enter : presenceDefinition.exit;
 
-      const animation = elementRef.current.animate(keyframes, {
-        fill: 'forwards',
+        const animation = animate(elementRef.current, keyframes, {
+          fill: 'forwards',
 
-        ...options,
-        ...(isReducedMotion() && { duration: 1 }),
-      });
+          ...options,
+          ...(isReducedMotion() && { duration: 1 }),
+        });
 
-      if (!visible && isFirstMount.current) {
-        // Heads up!
-        // .finish() is used there to skip animation on first mount, but apply animation styles immediately
-        animation.finish();
-        return;
-      }
+        if (!animation) {
+          return;
+        }
 
-      animationRef.current = animation;
-      animation.onfinish = visible ? onEnterFinish : onExitFinish;
+        if (!visible && isFirstMount) {
+          // Heads up!
+          // .finish() is used there to skip animation on first mount, but apply animation styles immediately
+          animation.finish();
+          return;
+        }
 
-      return () => {
-        animation.cancel();
-      };
-    }, [animationRef, isReducedMotion, onEnterFinish, onExitFinish, visible]);
+        animationRef.current = animation;
+        animation.onfinish = visible ? onEnterFinish : onExitFinish;
 
-    useIsomorphicLayoutEffect(() => {
-      isFirstMount.current = false;
-    }, []);
+        return () => {
+          animation.cancel();
+        };
+      },
+      // Excluding `isFirstMount` from deps to prevent re-triggering the animation on subsequent renders
+      // eslint-disable-next-line react-hooks/exhaustive-deps
+      [animationRef, isReducedMotion, onEnterFinish, onExitFinish, visible],
+    );
 
     if (mounted) {
       return React.cloneElement(child, { ref });