+++ packages/react-components/react-teaching-popover/src/components/TeachingPopoverTitle/useTeachingPopoverTitleStyles.styles.ts
@@ -27,7 +27,7 @@ const useStyles = makeStyles({
   },
   dismissButton: {
     position: 'relative',
-    ...shorthands.border(tokens.strokeWidthThin, 'solid', tokens.colorTransparentStroke),
+    border: `${tokens.strokeWidthThin} solid ${tokens.colorTransparentStroke}`,
     display: 'flex',
     alignItems: 'center',
     cursor: 'pointer',
@@ -38,8 +38,8 @@ const useStyles = makeStyles({
     borderBottomRightRadius: tokens.borderRadiusNone,
     borderRightStyle: 'none',
     ...createCustomFocusIndicatorStyle({
-      ...shorthands.outline(tokens.strokeWidthThick, 'solid', tokens.colorStrokeFocus2),
-      ...shorthands.borderRadius(tokens.borderRadiusMedium),
+      outline: `${tokens.strokeWidthThick} solid ${tokens.colorStrokeFocus2}`,
+      borderRadius: tokens.borderRadiusMedium,
       ...shorthands.borderColor('transparent'),
     }),
     marginInlineStart: 'auto',